[MR] 코드리뷰

See merge request sw_track/class_02_seoul/web_project_2/team16/healthcare-community!98

+++ server/src/routers/FoodRouter.ts
@@ -15,6 +15,23 @@ foodRouter.get('/', async (req, res, next) => {
   }
 });
 
+// 음식 검색
+foodRouter.get('/search', async (req, res, next) => {
+  try {
+    const keyword = req.query.foodName as string;
+
+    if (!keyword) {
+      throw new Error('검색 키워드가 반드시 필요합니다.');
+    }
+
+    const searchedFoodList = await foodService.searchFood(keyword);
+
+    res.status(200).json(searchedFoodList);
+  } catch (error) {
+    next(error);
+  }
+});
+
 // 음식 등록
 foodRouter.post('/register', async (req, res, next) => {
   try {
@@ -37,7 +54,7 @@ foodRouter.patch('/', async (req, res, next) => {
   try {
     if (is.emptyObject(req.body)) {
       throw new Error(
-        'headers의 Content-Type을 application/json으로 설정해주세요'
+        'headers의 Content-Type을 application/json으로 설정해주세요.'
       );
     }
 
@@ -45,9 +62,7 @@ foodRouter.patch('/', async (req, res, next) => {
       req.body;
 
     if (!foodId) {
-      throw new Error(
-        '수정을 위해서는 해당 음식의 ID(Object ID)가 필요합니다.'
-      );
+      throw new Error('해당 음식의 ID(Object ID)가 반드시 필요합니다.');
     }
 
     const toUpdateInfo = {
@@ -77,6 +92,10 @@ foodRouter.delete('/', async (req, res, next) => {
     }
 
     const { foodId } = req.body;
+
+    if (!foodId) {
+      throw new Error('해당 음식의 ID(Object ID)가 반드시 필요합니다.');
+    }
 
     const result = await foodService.deleteFood(foodId);
