+++ server/src/services/RoutineService.ts
@@ -1,5 +1,6 @@
 import { routineModel, RoutineModel } from '../db';
 import { userService } from './UserService';
+import { checkRoutine } from '../utils';
 
 export interface RoutineInfo {
   userId: string;
@@ -7,11 +8,17 @@ export interface RoutineInfo {
   routine_list: Array<Exercise>;
 }
 
-interface Exercise {
+export interface Exercise {
   name: string;
   count?: number;
   set?: number;
   weight?: number;
+}
+
+interface ErrorWithStatus {
+  status?: number;
+  message: string;
+  stack?: string;
 }
 
 class RoutineService {
@@ -27,19 +34,25 @@ class RoutineService {
     return result;
   }
 
+  async searchRoutine(userId: string, keyword: string) {
+    const routineList = await this.routineModel.searchRoutine(userId, keyword);
+    return routineList;
+  }
+
   async addRoutine(routineInfo: RoutineInfo) {
     const { userId, routine_name, routine_list } = routineInfo;
     // 검사 코드 추가
 
     const foundUser = await userService.findByUserId(userId);
     if (!foundUser) {
-      throw new Error('해당 유저는 존재하지 않습니다.');
+      throw new Error('해당 유저를 찾지 못했습니다.');
     }
+    const newRoutineList = checkRoutine(routine_list);
 
     const newRoutine = {
       userId,
       routine_name,
-      routine_list,
+      routine_list: newRoutineList,
     };
 
     const createdNewRoutine = await this.routineModel.create(newRoutine);
@@ -47,26 +60,42 @@ class RoutineService {
     return createdNewRoutine;
   }
 
-  async deleteRoutine(routineId: string) {
+  async deleteRoutine(userId: string, routineId: string) {
     const foundRoutine = await this.routineModel.findRoutineByObjectId(
       routineId
     );
 
     if (!foundRoutine) {
-      throw new Error('해당 루틴은 존재하지 않습니다.');
+      throw new Error('해당 루틴을 찾지 못했습니다.');
+    }
+
+    if (foundRoutine.userId !== userId) {
+      const error: ErrorWithStatus = new Error('작성자만 삭제할 수 있습니다.');
+      error.status = 403;
+      throw error;
     }
 
     const deletedRoutine = await this.routineModel.deleteByRoutineId(routineId);
     return deletedRoutine;
   }
 
-  async patchRoutine(routineId: string, toUpdateInfo: Partial<RoutineInfo>) {
+  async patchRoutine(
+    userId: string,
+    routineId: string,
+    toUpdateInfo: Partial<RoutineInfo>
+  ) {
     const foundRoutine = await this.routineModel.findRoutineByObjectId(
       routineId
     );
 
     if (!foundRoutine) {
-      throw new Error('해당 루틴은 존재하지 않습니다.');
+      throw new Error('해당 루틴을 찾지 못했습니다.');
+    }
+
+    if (foundRoutine.userId !== userId) {
+      const error: ErrorWithStatus = new Error('작성자만 수정할 수 있습니다.');
+      error.status = 403;
+      throw error;
     }
 
     const updatedRoutine = await this.routineModel.update(
